% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rustTiktoken.R
\name{tokens_count}
\alias{tokens_count}
\title{Rust tiktoken}
\usage{
tokens_count(text)
}
\value{
token counts evaluating in Rust (number, and NULL respectively).
}
\description{
Rust tiktoken
}
\examples{
tokens_count("This is a test string")
}
